missive: Extend resources interface
Add an ability to register callback if resource is not available, and be
called back once it is. Planned to be used on more critical paths only
like event enqueue - not on retryable once like upload.

Also, it does not make much sense to differentiate between disk and
memory resources - just initialize them differently.

BUG=b:258822636
TEST=FEATURES="test" emerge-$BOARD missive
TEST=missived_testrunner
--gtest_filter="*ResourceInterfaceTest.*" --gtest_repeat=10000

Change-Id: I4dac306cd2821cbb1dfa837b033ff51009b2040e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4089224
Commit-Queue: Leonid Baraz <lbaraz@chromium.org>
Tested-by: Leonid Baraz <lbaraz@chromium.org>
Reviewed-by: Hong Xu <xuhong@google.com>