shill: Fix the unit test of WiFiLinkStatisticsTest
WiFiLinkStatisticsTest alternates between calls to EXPECT_CALL() and
calls to the mock functions in DhcpFailure, NetworkValidationFailure and
DhcpNetworkValidationFailures, and therefore the behavior is undefined.
The reference is
https://google.github.io/googletest/gmock_for_dummies.html#using-mocks-in-tests
Running the unit test locally didn't reproduce the failure, therefore
this CL has not been proved to fix b/261667084.

Move calls to EXPECT_CALL() before calls to the mock functions, and
force gMock to verify and destroy the mock object before new calls to
EXPECT_CALL().

BUG=b:261667084
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: I546ac4d08f25eb22169e034a1605b403c4cb0fc3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4089949
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Commit-Queue: Kaidong Wang <kaidong@google.com>
Tested-by: Kaidong Wang <kaidong@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>