cryptohome: Merge USS and VK AuthFactors
Merge USS and VK AuthFactors on AuthSession start. So that when
user has a mix of different key backing stores during the
migration phase all the factors are merged. If migration is not
started for the user then this change doesn't have any effect
since we only expect a user has VaultKeyset or AuthFactor, but
not both.

BUG=b:258525620
TEST=FEATURES=test emerge-octopus cryptohome
USE=uss_migration FEATURES=test emerge-octopus cryptohome

Change-Id: I49e0b970eeff310fadcd767edb1a9b77639909d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4090374
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Betul Soysal <betuls@google.com>
Tested-by: Betul Soysal <betuls@google.com>
Reviewed-by: John Admanski <jadmanski@chromium.org>