libbrillo: Reset umask to default after running LVM commands
LVM commands run via liblvm2cmd set a umask(). While this is fine
if the LVM command is run as a process, the changes the default
umask if the command is run from the context of the same process.

Use a ScopedUmask to reset the umask() once we are done executing
the command.

BUG=b:261224746
TEST=CQ, manually check umask from chromeos_startup.sh

Change-Id: Id5ad0e1cdd0e2a0489fff4cae79af61837715f10
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4089678
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>