power: Remove custom timing code in keyboard backlight controller
We currently implement custom code in the `KeyboardBacklightController`
class to allow tests to simulate changes in time and timeouts. However,
because we now use a TaskEnvironment with a MOCK_TIME clock, we no
longer need to use ad-hoc methods, but can just simulate changes in the
global clock.

Remove the support code, and update tests to just modify the simulated
global clock directly.

BUG=b:260034799
TEST=`cros_workon_make --board=$BOARD --test chromeos-base/power_manager`

Change-Id: Id87ca6646c31ac765bf39972189e5ebf5abe8ad8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4091047
Reviewed-by: David Lattimore <dml@chromium.org>
Tested-by: David Greenaway <dgreenaway@google.com>
Commit-Queue: David Greenaway <dgreenaway@google.com>