power: Fix keyboard backlight timeout during fullscreen video playback
The current keyboard backlight code is designed to automatically turn
off the keyboard backlight:

1. For 30 seconds (the "keep on delay" preference) since the last user
   activity; or

2. For 3 seconds (the "keep on during video delay" preference) since the
   last user activity if a fullscreen video is playing.

The idea of the second item is that during full screen video playback we
want to turn off the backlight more quickly to avoid distractions.

However, the second part of this currently does not work correctly when
a video first starts playing, because we fail to update the "turn off
timer" when the video initially starts to play.

Existing tests don't catch this bug because we only test the feature
with the `turn_on_for_user_activity` feature set to false. In the field,
the setting is always set to true, however.

Update the backlight code to correctly set the timer, and update the
tests to exercise the feature with `turn_on_for_user_activity` set to
true. We no longer test the feature with the flag
`turn_on_for_user_activity` set to false, because an upcoming change
will remove that feature (b/260034799).

BUG=b:262289726
TEST=`cros_workon_make --board=$BOARD --test chromeos-base/power_manager`
TEST=Manually tested with fullscreen video playback, ensuring the
     timeout occurs after three seconds.

Change-Id: I0a24efe4f01e53b242778f2f3fcc84b87bc6d4bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4091048
Tested-by: David Greenaway <dgreenaway@google.com>
Reviewed-by: David Lattimore <dml@chromium.org>
Commit-Queue: David Greenaway <dgreenaway@google.com>