cryptohome: Fix crash on double CreatePersistentUser
We shouldn't crash if CreatePersistentUser is called for an
already existing user.

Specifically, the problem was hitting CHECK assertions when
AuthSession::OnUserCreated() is called twice in the "USS enabled"
mode (the CHECK is preventing creating UserSecretStash'es twice).
The problematic use case was when the user's root directory existed
(/home/.shadow/<shou>), but it contained no encrypted vault
subdirectories (vault/ or mount/).

The unit test coverage for the fix is an existing test that's now
becoming parameterized: run once with USS and without USS enabled.
We also had to invert the assertion in the test, since it doesn't
make sense that unnecessary CreatePersistentUser calls succeed.

BUG=b:261996750
TEST=new unit tests (cros_workon_make --test cryptohome)

Change-Id: Icb4966f79aab6b2242c9f5f361a7e18c98a53142
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4092325
Reviewed-by: John Admanski <jadmanski@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>