secanomalyd: adds |DEV| env var to upstart config
Adds a new env var to the upstart config file for secanomalyd. This
variable is imported and its value is used indirectly to determine
whether to pass the "--dev" flag to the secanomalyd executable.
The goal is to allow Tast tests to enable reporting in dev mode, as this
is needed to test that the UMA metrics are being emitted correctly.

BUG=b:257543432
TEST=secanomalyd starts without the --dev flag when the DEV variable is
not supplied, and starts with --dev when the variable is set to true.
TEST=secanomalyd start without the --dev flag when any value other than
"true" is supplied for the DEV variable, including "false".

Change-Id: I33e29cd3cb4b60ae5ac0efaea702842e051713f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4093137
Commit-Queue: Shervin Oloumi <enlightened@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Shervin Oloumi <enlightened@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>