shill: network: move PortalDetector to Network
This patch moves the PortalDetector member variable of the base Device
class to the Network class. The usage inside Device is still identical,
with the difference that the Device::OnNetworkValidationStart() and
OnNetworkValidationStop() methods are now defined on the Network's
EventHandler class.

There is no behavior change in this patch.

BUG=b:248028325
BUG=b:265607618
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: I15a2b79cfb2dfe9da54e51e8b94ea4438e157a34
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4141268
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Hugo Benichi <hugobenichi@google.com>