shill: portal: simplifying handling of portal state results
This patch reorganizes the code in Device::PortalDetectorCallback that
reacts to the Service connection state derived from the portal detection
result. This prepares the introduction of a new enum for the
return value of PortalDetector::GetConnectionState() and for
encapulating the network validation retry loop inside PortalDetector.

BUG=b:248028325
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: If0d8f74d0a6454a4f4962be4b4c188266a8028ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4141266
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Tested-by: Hugo Benichi <hugobenichi@google.com>