vm_tools: update concierge to create foz_db_list file for crosvm
|foz-db-list-path| is the newly introduced variable for crosvm.
This is used to set the path to a text file, which is a list of
relative paths to RO single file shader caches.

https://crrev.com/c/4096398 adds the new variable for crosvm,
but crosvm does not utilize this for now.

This feature is only available for Borealis, for now.

BUG=b:262181746
TEST=In DUT: `vmc launch borealis`

Disallow-Recycled-Builds: test-failures
Cq-Depend: chromium:4173588
Change-Id: I1560552c7050993244dd4f8de70131e228e2f829
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4096059
Tested-by: Max Lee <endlesspring@google.com>
Commit-Queue: Max Lee <endlesspring@google.com>
Reviewed-by: Nic Hollingum <hollingum@google.com>