shill: Emit BT stack info in WiFiLinkQualityReport
BT can have an impact on WiFi (e.g. CoEx) so it is useful to gather some
information about BT (is it on? is it scanning?) as part of the WiFi
link quality report. Design proposal is available at
go/cros-wifi-bt-info-in-wifi-metrics.

Query which BT stack is in use (Floss or BlueZ) and add it to the
WiFiLinkQualityReport structured event.

BUG=b:253297825
TEST=cros-workon-${BOARD} start shill
TEST=FEATURES=test emerge-${BOARD} shill
TEST=cros deploy ${DUT} shill; restart shill
TEST=observe in the log the BT stack reported in the
WiFiLinkQualityReport event switches between Floss and BlueZ when the
bluetooth-use-floss Chrome flag is flipped.
TEST=stop btmanagerd, observe that the "unknown" BT stack is reported.
TEST=start btmanagerd, disable Floss, observe that the BlueZ BT stack is
reported.
TEST=start btmanagerd, enabled Floss, observe that the Floss BT stack is
reported.

TEST=USE=-floss" emerge-${BOARD} shill
TEST=cros deploy ${DUT} shill; restart shill
TEST=observe in the debug log that the BT stack in use is always
"unknown"

Change-Id: I0f187228ab49c4cd3ab4ecc56c2a7c3d2438ac96
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4098608
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: David Ruth <druth@chromium.org>
Commit-Queue: Nicolas Norvez <norvez@chromium.org>