typecd: Add wrapper for fetching browser setting state
In preparation for the move to relying on cached setting state,
introduce a wrapper for the IsPeripheralDataAccessEnabled() method,
which calls that function and updates the local copy of the setting.

Call this new wrapper during daemon init to ensure that the cached
setting state is up to date on init.

Once we have completely moved to using the cached state,
IsPeripheralDataAccessEnabled() will be folded into its wrapper.

BUG=b:260017469
TEST=All unit tests pass.

Change-Id: Icc7a8187a6fd5e0b9bdfa6a3846ac3b6fb9440e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4098612
Commit-Queue: Prashant Malani <pmalani@chromium.org>
Reviewed-by: Jameson Thies <jthies@google.com>
Tested-by: Prashant Malani <pmalani@chromium.org>