diagnostics: Add audio jack event
This CL supports audio jack event for |EventObserver| interface. With
this, our clients can register it and monitor audio jack related events.
One of the use case is diagnostics, the diagnostic app can ask user to
plug in or unplug the audio jack and see if we can report the events
correctly.

The first use case will be OEM PWA (b/255517102). COST team will wrap
our API and expose it as an extension API quarterly. That is, the caller
code is not covered in this CL.

Besides, this CL fixes the bad behavior that cros-health-tool won't
terminate when the observer is gone.

BUG=b:255517102, b:260692532
TEST=FEATURES=test emerge-hatch diagnostics
TEST=1. (DUT) cros-health-tool event --category=audio_jack
        --length_seconds=600
     2. Plug in a 3.5mm devices.
     3. It detects the event successfully.
TEST=1. (DUT) cros-health-tool event --category=audio_jack
        --length_seconds=600
     2. Kill the delegate process.
     3. The cros-health-tool terminates immediately.

Change-Id: I424ce7b2828b12fb8d3972f63396b6d330273376
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4121797
Reviewed-by: Sam McNally <sammc@chromium.org>
Tested-by: Kerker Yang <kerker@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Wei-Luan Wang <weiluanwang@google.com>
Commit-Queue: Kerker Yang <kerker@chromium.org>