installer: Fix postinst bios_type UMA
`bios_type` as passed to RunPostInstall is often `kUnknown`, with the
expectation that it will be detected in `ConfigureInstall`. Sending the
`Installer.Postinstall.BiosType` metric before detection is not useful.

Update when and how we send the metric to capture the detected bios type
or whatever we get, even if there's a failure.

As an aside: the fact that an incorrect `bios_type` sticks around in
`RunPostInstall`, providing a small footgun, feels like something worth
fixing. I took a look at pulling the detection of bios type up out of
`ConfigureInstall`, but I think it was worse in other ways. A more
extensive change could remove this risk, but that doesn't quite feel
warranted here.

BUG=b:235257917
TEST=FEATURES=test emerge-reven chromeos-installer

Change-Id: I93ff0f0b645d4f548f2229b5ad9ac1bcd3d75b89
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4101840
Commit-Queue: Ted Brandston <tbrandston@google.com>
Tested-by: Ted Brandston <tbrandston@google.com>
Reviewed-by: Nicholas Bishop <nicholasbishop@google.com>
Auto-Submit: Ted Brandston <tbrandston@google.com>