libipp: Remove enum AttrState
This CL removes the enum AttrState and related methods. The type
AttrState was used to store out-of-band tags. Now, these values can be
stored with the enum ValueTag. Each attribute had a AttrState variable
that was stored in a map inside the Collection class. This value was
moved to def_.ipp_type inside the Attribute class.
This CL also sorts out the logic related to an attribute's state:
 * when an attribute has out-of-band tag (IsOutOfBand(Tag())), the
   attribute has always no values (Size() == 0);
 * when an attribute has non-out-of-band tag (!IsOutOfBand(Tag())),
   the attribute has always at least one value (Size() > 0).

BUG=b:216051238
TEST=FEATURES=test emerge-amd64-generic libipp

Change-Id: I281c63add2bbe86060533a93ccc9ce3994cbe780
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4104220
Commit-Queue: Edward Hill <ecgh@chromium.org>
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Piotr Pawliczek <pawliczek@chromium.org>