login: write-machine-info: Log errors when writing the machine info
It's postulated that on newer ARM boards we're incorrectly falling
back to a bogus "WPSW_BOOT" value when writing the machine info. In
<https://crrev.com/c/4104708> we're now printing to stderr when we
fall back, but that just goes to stderr. In order to see anything in
device logs, we actually have to capture this stderr and send it to
logs. Let's do that here.

BUG=b:249498455, b:254337014
TEST=Force a failure and see the log

Change-Id: I11e193f274a2310451c80b6213e04f61b882cf71
Disallow-Recycled-Builds: test-failures
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4103931
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Auto-Submit: Douglas Anderson <dianders@chromium.org>