camera: None effect bypasses expensive section of EffectsSteamManipulator
The none effect still goes through the conversion from YUV to RGBA and
back again as well as the buffer copy in libcros_ml_core_internal.so.
This is unnecessary work.

This change skips the these heavy lifting sections when the effect is
detected as being none.

BUG=b:249868506
TEST=emerge-brya cros-camera-libs; manual test on device

Change-Id: I131c7e8cd2ce2a1ae1a0a031b7b4d1e58355b98a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4074920
Tested-by: Tom Shafron <shafron@google.com>
Commit-Queue: Tom Shafron <shafron@google.com>
Reviewed-by: Jim Pollock <jmpollock@chromium.org>
Reviewed-by: Ren-Pei Zeng <kamesan@chromium.org>