oobe_config: Move UMA metrics to new metrics folder
We are planning to improve the metrics for enterprise
rollback to ensure they do not get lost when the device
is powerwashed. It is required to create a new library
to handle the metric process.

Move existing metric files to its own folder. This folder
will contain all the logic related to metrics to keep the
code organized. Rename the existing metric files to make
a clear distinction between UMA metrics and the structured
metrics that we are planning to add.

BUG=b:261850979
TEST=FEATURES=test emerge-amd64-generic oobe_config

Change-Id: I421892cc2ec6bc81dcfadffad3513143a9f6c8ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4106120
Reviewed-by: Miriam Polzer <mpolzer@google.com>
Tested-by: Cristina Guerrero <crisguerrero@chromium.org>
Commit-Queue: Cristina Guerrero <crisguerrero@chromium.org>