cryptohome: Ignore malformed dmcrypt mount paths
When doing low-level file path manipulation in the code that
detects stale dmcrypt mounts, skip entries that would lead to
creating nonsensical paths with "/../".

This should probably never happen in practice, but it seems it
won't harm to check for this explicitly. Also it fixes spurious
bug reports from userdataauth_fuzzer as it treats any path with
"/../" as a potential security issue.

BUG=b:254864841
TEST=userdataauth_fuzzer doesn't hit debug assertions on "!ReferencesParent()"

Change-Id: Ib059d8c05c5f907a1a63f109b7c545b0742b26f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4106573
Reviewed-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Daniil Lunev <dlunev@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>