chromeos-common-script: Check found root device type matches expectation
When cros_config enforces a root device type, reject found root device
that does not match the type. It can be overridden by specifying the
--dst argument.

On device that can have NVMe and eMMC device, enforce omitting the eMMC
device when cros_config insists the root device type is NVMe. It can
happen when the NVMe device is not present or broken.

BUG=b:229020025
TEST=Line by line on snappy
Using herobrine:
Set
DEFAULT_ROOTDEV="/sys/devices/platform/soc*/1c08000.pci/pci0001:00/0001:00:00.0/0001:01:00.0/nvme/nvme0/nvme0n1
/sys/devices/platform/soc*/7c4000.sdhci/mmc_host/mmc*/mmc*:*/block/mmcblk*"
Check that when |rootdev| is set to /dev/nvme0n1 or /dev/mmcblk1, a non
null |dev| is returned when |exp_rootdev_type| is empty or unknown.

When |exp_rootdev_type| is set to EMMC, only when |rootdev| set to
/dev/mmcblk1 a non null |dev| is returned.
When |exp_rootdev_type| is set to NVME, only when |rootdev| set to
/dev/nvme0n1 a non null |dev| is returned.

Change-Id: I4759272511427ec4ba900260d610fbd58764cbeb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4108609
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>