typecd: add horizontal and vertical position to physical location
Currently there is a panel property in the port class of typecd that
allows the daemon to know which panel each USB-C port is located on.
But for boards with multiple ports on the same panel, there is currently
no way to differentiate them. This CL adds horizontal and physical
position to the physical location data in the port class to know where
on each panel different ports are located.

BUG=b:262609382
TEST=passing all unit tests including updated unit tests which check
that typecd correctly readers the horizontal and vertical position sysfs
paths.

Change-Id: I9fa846949c97bf8137c505e43c7892aac6998bca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4109468
Tested-by: Jameson Thies <jthies@google.com>
Reviewed-by: Prashant Malani <pmalani@chromium.org>
Reviewed-by: Won Chung <wonchung@google.com>
Commit-Queue: Jameson Thies <jthies@google.com>