crash: Include iommu_group information in SMMU faults
This makes it sometimes possible to correlate a context bank number from
the fault (cb=N) to the device that it is associated with.

Example:
 ==iommu groups==
 /sys/kernel/iommu_groups/7/devices/a600000.usb
 DMA
 0x0000000008000000 0x00000000080fffff msi

This could mean context bank 7 is used for usb and the iommu domain is a
DMA domain with an IOVA region of 0x8000000 to 0x80fffff.

BUG=None
TEST=FEATURES="test" emerge-${BOARD} crash-reporter
TEST=Simulate crash with lkdtm and check output of log.gz in crash spool

Change-Id: I7ec838262cd29678a88bbc3c1d4454af16e342b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4109392
Tested-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Rob Clark <robdclark@chromium.org>