shill: wifi: stop network when connecting to a new Service
See $BUG for details, we aren't currently resetting the L3 state when
connecting to a new network. This is generally okay because we'll end up
reconfiguring after we connect, but when we receive lingering dhcpcd
events on the old network while connecting to the new one, we'll end up
bypassing all L3 configuration for the new network. Explicitly reset the
L3 state in this patch.

BUG=b:261918958
TEST=FEATURES=test emerge-guybrush shill
TEST=Deploy to a dewatt device, `tast run ${DUT} wifi.ProfileBasic` 15
times, no failures anymore

Change-Id: Ic8e3787e6a59c42ee831aa372fdd140c10891187
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4112608
Commit-Queue: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Matthew Wang <matthewmwang@chromium.org>