shill: vpn: Start charon also when failed to read pid file
In the current implementation, we will not start charon if we failed to
get the status of whether charon is running, e.g., charon pid file
exists but its content is invalid. This patch change the logic to a more
optimistic way so that we'll start charon unless we're sure that there
is another charon running. We can get an error if charon quits itself
afterwards anyway.

BUG=b:261666421
TEST=FEATURES="test" emerge-${BOARD} shill
TEST=../platform/tast-tests/tools/repeat-test.sh -s -r 30 -- $DUT \
  network.VPNDNS*

Change-Id: I866be1feb68b2e4d8331420be4656436a999ee75
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4113543
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>