shill: Parse phy concurrency in WiFiPhy
Adds the function ParseConcurrency into WiFiPhy, which parses
concurrency cababilites from nl80211 new wiphy message. These
capabilities are stored in a format that directly mirrors the
format in which they are received in the new wiphy message.

This change does not include the WiFiPhy class taking any
responsibility for concurrency decisions. It only includes
pulling in and parsing the concurrency info and exposing it.

BUG=b:222499165
TEST=FEATURES=test emerge-hatch shill

Change-Id: I1ba4be0b87f2083b93cd77d9863cdee068dda718
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4121357
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Tested-by: Kevin Lund <kglund@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Jintao Lin <jintaolin@chromium.org>