foomatic_shell: Detect and report violation of CPU time limit
This patch introduces a new return code 126 that is supposed to be
returned by a child process when maximum CPU time is reached. Also, the
total CPU time of children processes will be reported to new UMA
metrics. Moreover, the limit of CPU time is increased to 15 minutes.

BUG=b:182409731
TEST=FEATURES=test emerge-amd64-generic foomatic_shell

Cq-Depend: chromium:4114668
Change-Id: I9c5e94b22e66ab0192384c7f0b71d329e1307f48
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4115896
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Piotr Pawliczek <pawliczek@chromium.org>
Commit-Queue: Piotr Pawliczek <pawliczek@chromium.org>