shill: cellular: Add the ip-type as a differentiator for APNs
Currently, the IP type is not being taken into account when comparing
different APNs. This causes the IP type of a modem APN to be overridden
by the modb APN IP type when the APNs are almost equal.
Adding the IP type in the key would allow the modem APN to be used
as it is, and if the connection fails, the modb APN will be used.

BUG=None
TEST=Inject a modem profile in ModemManager with a combination of values
and verify the different behaviors in the APN UI.

Change-Id: If8693c579b0ba92659039e9880ecc9319db70d99
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4118965
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>