shill: cellular: Remove warning message for MCCMNC not found
Because AppendToCandidatesByMCCMNC is called with a 5 digit and 6 digit
MCCMCN extracted from the IMSI, the function always prints a warning
message that confuses people when reading logs. Improve the logging of
this warning so it is only printed when none of the MCCMNCs are found.

BUG=None
TEST=Verify that the message is not printed.

Change-Id: Icffe71cab8d53d8efd944daae97c918cc69b759a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4114239
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Auto-Submit: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Madhav <madhavadas@google.com>