cryptohome: VaultKeyset should see |reset_secret| in KeyBlobs
Before adding a pinweaver VaultKeyset we double check that the
VaultKeyset has a reset_seed. reset_seed is not stored in a pin
VaultKeyset but this check ensures that we don't create a
pinweaver VaultKeyset that doesn't have a reset mechanism.
This is ensured since reset_seed must exist on a password
VaultKeyset to create a pinweaver VaultKeyset, and since it is
passed to new created pinwevaer VaultKeyset, just to check its
existence before persisting the VaultKeyset.

However, during the migration of a pinweaver VaultKeyset to USS
we override the existing pinweaver VaultKeyset with a migrated
VaultKeyset of the same field values. But reset_seed and
reset_secret fields are not passed to the newly created keyset.
The missing reset_secret and reset_seed fields result in not
persisting the migrated VaultKeyset, that means VaultKeyset
migration cannot be completed. So VaultKeyset::EncryptEx()
should see the existence of the reset_secret in the KeyBlobs
to proceed with encryption and persisting of  a pinweaver
VaultKeyset. It doesn't use reset_secret or store reset_secret,
but this is needed as a check mechanism to ensure that the
created pinweaver node has a reset_secret.

This CL also adds some useful debug info.

BUG=b:258465405
TEST=tast run DUT cryptohome* hwsec*
TEST=tast run DUT cryptohome.UssMigrationPinPassword

Change-Id: I4bca9d0ecc52f082640354424fe6783e42851cf0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4117818
Reviewed-by: Thomas Cedeno <thomascedeno@google.com>
Commit-Queue: Betul Soysal <betuls@google.com>
Tested-by: Betul Soysal <betuls@google.com>