cryptohome: Extract auth factor loading
This change extracts the auth factor loading logic in the AuthSession
Initialize into a separate function. This is part of an intended larger
sequence of changes to externalize a lot of the AuthSession setup in
order to make testing with AuthSession easier.

This initial change just extracts the loading function without trying to
rearrange the AuthSession construction yet.

BUG=b:263259501
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Ib534647c1a04ee72b8089069725dfc083261de33
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4119249
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>