cryptohome: Report empty backing store metrics
The current reporitng of the backing store metrics misses the early
exit code path in AuthSession initialization. This fixes that gap.

In practice it's not a big deal, because in this case you'll always just
get an "empty" backing store, and we mostly just care about VK vs USS vs
mixed stores, but it seems better to have consistency.

BUG=b:258711982
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Ie5a6ac161820b01f219554b1d40b3227dff6f214
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4119253
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>