ml_benchmark: update model zoo with new models, move to google storage
This CL fixes the convolution benchmarks that contained NaN's in some of
the weights.

This also introduces the mobilenet models are a proxy benchmark for
image search.

This also moves the storage of the model files out of git and into
google storage.

BUG=b:263313506
TEST=None

Cq-Depend: chromium:4113504
Change-Id: I708f94a99d6ed5b9da2dfb9aa50e8a3fad0c4a7d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4113556
Commit-Queue: Jim Pollock <jmpollock@chromium.org>
Reviewed-by: Michael Olney <mpolney@google.com>
Tested-by: Jim Pollock <jmpollock@chromium.org>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>