power: Replace the OneShot with RepeatingTimer
This change will make ReadBatteryStatus to be run every 10 minutes.

This change is a partial fix.
Unfortunately, we have not managed to reproduce the issue.
ReadBatteryStatus method clears the vector, that for a very small
subset of Chromebooks in the global fleet reaches over 1000+ FDs.

Replacing a OneShotTimer with a RepeatingTimer intends to reduce a
chance that a single client hits 1000+ FDs in the vector.
ReadBatteryStatus is called via D-Bus from Chrome ash with irregular
frequency (which I have not measured or verified; but frequency small
enough that I have not observed it in local testing for hours).

BUG=b:207716926
TEST=emerge-strongbad chromeos-base/power_manager; cros deploy and run
on DUT

Change-Id: If557ff6e3e4406c54849a86944822c0123000e58
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4118457
Tested-by: Jakub Młynarczyk <jakubm@chromium.org>
Reviewed-by: David Lattimore <dml@chromium.org>
Commit-Queue: Jakub Młynarczyk <jakubm@chromium.org>
Reviewed-by: David Greenaway <dgreenaway@google.com>
Reviewed-by: Mengqi Guo <mqg@chromium.org>