camera: remove VLOGF_ENTER() and VLOGF_EXIT()
Since we now have good support for event tracing, we can remove all the
VLOGF_ENTER and VLOGF_EXIT logs and replace them with trace events if
needed. This CL removes all the verbose enter and exit logs.

The binary sizes are smaller with the logs removed. Just to name a few
from offline testing:

With the VLOGs:

2548592    /build/brya/usr/lib64/libcros_camera.so
 476768    /build/brya/usr/bin/cros_camera_service

and without:

2523088    /build/brya/usr/lib64/libcros_camera.so
 456736    /build/brya/usr/bin/cros_camera_service

BUG=b:212231270
TEST=CQ

Change-Id: I8a4a5edeb0b183cfefc4ab4ebd0a333f1d4a1823
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4116268
Tested-by: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Ren-Pei Zeng <kamesan@chromium.org>
Commit-Queue: Ricky Liang <jcliang@chromium.org>