cryptohome: Drop StartAuthSession.key_label_data
This removes the support that populates the key_label_data field in the
StartAuthSession reply.

This code is only used on a "!use-auth-factors" path in the client code,
which is in the process of being removed. We obviously can't remove the
field itself from the proto definition until all such uses (on both the
client and server side) are ripped out by we can more aggressively drop
the references on the server side.

BUG=b:261030151
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Idb89b661ed69c2a2aad143a191b942288226852f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4119739
Reviewed-by: Betul Soysal <betuls@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: Betul Soysal <betuls@google.com>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>