libipp: Replace raw pointers in Attribute API by iterators
The API of Attribute has a method GetCollection() that returns raw
pointer to a Collection. This CL marks this method as deprecated and
introduces new method Colls() that returns CollsView object. This
object provides an interface allowing to iterate over Collections, as
well as, operator[] and methods size(), empty(). This interface is also
used in the Frame API by a new method Groups().

BUG=b:216051238
TEST=FEATURES=test emerge-amd64-generic libipp

Change-Id: Iac39d6bdf507f558b7979fe5763d32e17536d58c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4121309
Commit-Queue: Piotr Pawliczek <pawliczek@chromium.org>
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Piotr Pawliczek <pawliczek@chromium.org>