shill: handle multiple BT adapters
If more than 1 BT adapter is enabled, do not directly query the BT
adapters. Instead, ask the BT manager which adapter is the default one,
and then query the state of that adapter.

BUG=b:253297825
TEST=cros-workon-${BOARD} start shill
TEST=FEATURES=test emerge-${BOARD} shill
TEST=emerge-${BOARD} shill
TEST=cros deploy ${DUT} shill; restart shill
TEST=observe in the debug log that we're querying the enabled adapter.
The case of multiple adapters has *not* been tested (lack of hardware)

TEST=FEATURES=test USE=-floss" emerge-${BOARD} shill

Change-Id: I76230995ad2a075e7149a5fe7dd16647d691715b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4119700
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: David Ruth <druth@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>