cryptohome: data_migrator: Add an option not to ignore EIO failure
cryptohome::data_migrator::MigrationHelper currently ignores file open
failures with EIO. Although the configuration was suitable for dircrypto
migration, it isn't for ARCVM /data migration. This CL thus adds an
option to ignore EIO failure to the delegate of MigrationHelper so that
the EIO failures are only ignored in dircrypto migration.

This CL also adds a comment to describe the historical reason of the
name of the file to mark the start of the migration.

BUG=b:258402655
TEST=FEATURES=test emerge-hatch cryptohome arcvm-data-migrator
TEST=Manually perform ARCVM /data migration.
TEST=tast run <DUT> cryptohome.VaultMigration.fscrypt_v1
TEST=tast run <DUT> cryptohome.VaultMigration.fscrypt_v2

Change-Id: Id75c55a68fe36c4dfc09fee1ca99cca82651fc37
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4202649
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Tested-by: Momoko Hattori <momohatt@chromium.org>
Commit-Queue: Momoko Hattori <momohatt@chromium.org>