libchromeos-ui: Disable Floss for boards that do not support it.
Currently, if the Floss flag is enabled and Floss does not exist on a
given 2GB rootfs board, Bluetooth is disabled entirely.

To fix this, we prevent users from enabling Floss in the first place
by override enabling/disabling Floss in the Chromium UI if the USE
flag is present/absent. The presence of the USE flag is indicated by
an explicit command line disable when Chrome is built.

BUG=b:228902194
TEST=manual

Cq-Depend: chromium:4128738
Change-Id: I6b1a83f20ff2b1c5e955439ca7d108339b8a9347
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4127877
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Sarvesh Kalwit <sarveshkalwit@google.com>