ml: Add madvise and mprotect syscalls
At present, many of the ml models cannot call either madvise or
mprotect. This prevents other allocators, such as scudo or
partition alloc, from replacing the glibc allocator which doesn't use
these syscalls.

mprotect in particular is a fairly dangerous syscall to add, so we
need to be careful that there's never a situation where both PROT_EXEC
and PROT_WRITE are set with mprotect.

I'm not sure which policy actually needs to be edited. The error
is for ml_service, and I have no further information than that. But
the policies seem to be a mish-mash of allowing and disallowing these
syscalls anyways, so CL adds some uniformity.

BUG=b:263994163
TEST=CQ

Change-Id: I817497ca88dc5c93ab153626c55799c6b12f1af7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4129822
Reviewed-by: Allen Webb <allenwebb@google.com>
Reviewed-by: Andrew Moylan <amoylan@chromium.org>
Tested-by: Jordan Abrahams-Whitehead <ajordanr@google.com>
Commit-Queue: Jordan Abrahams-Whitehead <ajordanr@google.com>