federated: rename client_version to brella_lib_version
The word "client" was a bit overloaded. In most cases it refers to
projects that utilize federated service for analytics or training.
While in "client_version" it refers to the ChromeOS device version
that can be recognized by the server side to determine whether the
device is eligible for a given task.

This CL changes the latter case to "brella_lib_version" for better
readability.

BUG=b:239623649
TEST=cros_workon_make --board=betty --test federated-service

Change-Id: Id0195a7c13c77a07fde04c56d16dd41eb390f436
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4134325
Commit-Queue: Xinglong Luan <alanlxl@chromium.org>
Tested-by: Xinglong Luan <alanlxl@chromium.org>
Reviewed-by: Andrew Moylan <amoylan@chromium.org>